sprint_3 исправно работающий проект
