+++ src/main/java/scooter/Orders.java
@@ -1,5 +1,7 @@
 package scooter;
+
 import org.apache.commons.lang3.RandomStringUtils;
+
 import java.util.List;
 
 public class Orders {
@@ -14,36 +16,35 @@ public class Orders {
     public String comment;
     public List<String> color;
 
-
-    public Orders (String firstName, String lastName, String address, String metroStation, String phone, String rentTime, String deliveryDate, String comment) {
-            this.firstName = firstName;
-            this.lastName = lastName;
-            this.address = address;
-            this.metroStation = metroStation;
-            this.phone = phone;
-            this.rentTime = rentTime;
-            this.deliveryDate = deliveryDate;
-            this.comment = comment;
-        }
-
-        public static Orders getRandom() {
-            final String firstName = RandomStringUtils.randomAlphabetic(10);
-            final String lastName = RandomStringUtils.randomAlphabetic(10);
-            final String address = RandomStringUtils.randomAlphabetic(10);
-            final String metroStation = RandomStringUtils.randomAlphabetic(10);
-            final String phone = RandomStringUtils.randomNumeric(11);
-            final String rentTime = RandomStringUtils.randomNumeric(1);
-            final String deliveryDate = "2022-02-15";
-            final String comment = RandomStringUtils.randomAlphabetic(10);
-            return new Orders(firstName, lastName, address, metroStation, phone, rentTime, deliveryDate, comment);
-        }
-
-        public Orders setColor(List<String> color) {
-            getRandom();
-            this.color = color;
-            return this;
-        }
-
+    public Orders(String firstName, String lastName, String address, String metroStation, String phone, String rentTime, String deliveryDate, String comment) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.address = address;
+        this.metroStation = metroStation;
+        this.phone = phone;
+        this.rentTime = rentTime;
+        this.deliveryDate = deliveryDate;
+        this.comment = comment;
     }
+
+    public static Orders getRandom() {
+        final String firstName = RandomStringUtils.randomAlphabetic(10);
+        final String lastName = RandomStringUtils.randomAlphabetic(10);
+        final String address = RandomStringUtils.randomAlphabetic(10);
+        final String metroStation = RandomStringUtils.randomAlphabetic(10);
+        final String phone = RandomStringUtils.randomNumeric(11);
+        final String rentTime = RandomStringUtils.randomNumeric(1);
+        final String deliveryDate = "2022-02-15";
+        final String comment = RandomStringUtils.randomAlphabetic(10);
+        return new Orders(firstName, lastName, address, metroStation, phone, rentTime, deliveryDate, comment);
+    }
+
+    public Orders setColor(List<String> color) {
+        getRandom();
+        this.color = color;
+        return this;
+    }
+
+}
