+++ src/main/java/scooter/RestAssuresService.java
@@ -1,4 +1,5 @@
 package scooter;
+
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
