+++ src/test/java/scootertest/ColorOrderTest.java
@@ -1,11 +1,14 @@
 package scootertest;
+
 import io.restassured.response.ValidatableResponse;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import scooter.Orders;
 import scooter.OrderService;
+
 import java.util.List;
+
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
