+++ src/test/java/scootertest/ListOderTest.java
@@ -1,9 +1,12 @@
-package scooterTest;
+package scootertest;
+
 import io.restassured.response.ValidatableResponse;
 import org.junit.Test;
 import scooter.OrderService;
+
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
+import static org.junit.Assert.assertEquals;
 
 public class ListOderTest {
 
@@ -15,6 +18,8 @@ public class ListOderTest {
         public void orderListTest() {
             ValidatableResponse response = orderService.getOrderList();
             response.assertThat().body("orders.size()", is(not(0)));
+            int statusCode = response.extract().statusCode();
+            assertEquals("Status code is wrong.", 200, statusCode);
         }
 
     }
