+++ src/test/java/scootertest/OrderTest.java
@@ -1,33 +1,35 @@
 package scootertest;
+
 import io.restassured.response.ValidatableResponse;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import scooter.Orders;
 import scooter.OrderService;
+
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 
 public class OrderTest {
 
-        private Orders order;
-        private OrderService orderService;
-        private int orderTrack;
+    private Orders order;
+    private OrderService orderService;
+    private int orderTrack;
 
-        @Before
-        public void setUp() {
-            orderService = new OrderService();
-            order = Orders.getRandom();
-        }
+    @Before
+    public void setUp() {
+        orderService = new OrderService();
+        order = Orders.getRandom();
+    }
 
-        @Test
-        public void createRequestReturnsSuccessCodeAndIdTest() {
-            ValidatableResponse response = orderService.create(order);
-            int statusCode = response.extract().statusCode();
-            orderTrack = response.extract().path("track");
-            assertEquals("Status code is wrong.", 201, statusCode);
-            assertNotEquals("Id is not returned.", 0, orderTrack);
-        }
+    @Test
+    public void createRequestReturnsSuccessCodeAndIdTest() {
+        ValidatableResponse response = orderService.create(order);
+        int statusCode = response.extract().statusCode();
+        orderTrack = response.extract().path("track");
+        assertEquals("Status code is wrong.", 201, statusCode);
+        assertNotEquals("Id is not returned.", 0, orderTrack);
+    }
 
 
     @After
@@ -35,5 +37,5 @@ public class OrderTest {
         orderService.cancel(orderTrack);
     }
 
-    }
+}
